Move fmn_fasshim out of consumers to avoid circular import
