+++ fmn/consumer/digests.py
@@ -21,11 +21,11 @@ import fedmsg_meta_fedora_infrastructure
 
 from twisted.internet import reactor, task
 
+from fmn import fmn_fasshim
 import fmn.lib
 import fmn.rules.utils
 import fmn.consumer.producer as fmn_producers
 
-import fmn.consumer.fmn_fasshim
 from fedmsg_meta_fedora_infrastructure import fasshim
 
 log = logging.getLogger("fmn")
@@ -36,20 +36,20 @@ fedmsg.meta.make_processors(**CONFIG)
 DB_URI = CONFIG.get('fmn.sqlalchemy.uri', None)
 session = fmn.lib.models.init(DB_URI)
 
-fmn.consumer.fmn_fasshim.make_fas_cache(**CONFIG)
+fmn_fasshim.make_fas_cache(**CONFIG)
 # Monkey patch fedmsg_meta modules
-fasshim.nick2fas = fmn.consumer.fmn_fasshim.nick2fas
-fasshim.email2fas = fmn.consumer.fmn_fasshim.email2fas
+fasshim.nick2fas = fmn_fasshim.nick2fas
+fasshim.email2fas = fmn_fasshim.email2fas
 fedmsg_meta_fedora_infrastructure.supybot.nick2fas = \
-    fmn.consumer.fmn_fasshim.nick2fas
+    fmn_fasshim.nick2fas
 fedmsg_meta_fedora_infrastructure.anitya.email2fas = \
-    fmn.consumer.fmn_fasshim.email2fas
+    fmn_fasshim.email2fas
 fedmsg_meta_fedora_infrastructure.bz.email2fas = \
-    fmn.consumer.fmn_fasshim.email2fas
+    fmn_fasshim.email2fas
 fedmsg_meta_fedora_infrastructure.mailman3.email2fas = \
-    fmn.consumer.fmn_fasshim.email2fas
+    fmn_fasshim.email2fas
 fedmsg_meta_fedora_infrastructure.pagure.email2fas = \
-    fmn.consumer.fmn_fasshim.email2fas
+    fmn_fasshim.email2fas
 
 CNT = 0
